+++ src/Views/BucketPanel/BucketDetail.tsx
@@ -5,6 +5,8 @@ import BucketCard, {getHistory} from "../../Components/Bucket/BucketCard";
 // @ts-ignore
 import prettierBytes from "prettier-bytes";
 import {Table, Typography} from "antd";
+import {DeleteOutlined} from "@ant-design/icons";
+import RemoveConfirmationByName from "../../Components/RemoveConfirmationByName";
 
 interface Props {
     client: Client;
@@ -18,6 +20,7 @@ export default function BucketDetail(props: Readonly<Props>) {
 
     const [info, setInfo] = useState<BucketInfo>();
     const [entries, setEntries] = useState<EntryInfo[]>([]);
+    const [entryToRemove, setEntryToRemove] = useState<string>("");
 
     const getEntries = async () => {
         try {
@@ -31,11 +34,25 @@ export default function BucketDetail(props: Readonly<Props>) {
 
     };
 
+    const removeEntry = async (name: string) => {
+        if (!info) {
+            console.error("No bucket info");
+            return;
+        }
+
+        const {client} = props;
+        const bucket: Bucket = await client.getBucket(info.name);
+        await bucket.removeEntry(name);
+        setEntryToRemove("");
+        getEntries().then();
+    };
+
     useEffect(() => {
         getEntries().then();
         const interval = setInterval(() => getEntries(), 5000);
         return () => clearInterval(interval);
     }, []);
+
 
     const data = entries.map(entry => {
         const printIsoDate = (timestamp: bigint) => entry.recordCount !== 0n ?
@@ -48,7 +65,7 @@ export default function BucketDetail(props: Readonly<Props>) {
             size: prettierBytes(Number(entry.size)),
             history: entry.recordCount !== 0n ? getHistory(entry) : "---",
             oldestRecord: printIsoDate(entry.oldestRecord),
-            latestRecord: printIsoDate(entry.latestRecord),
+            latestRecord: printIsoDate(entry.latestRecord)
         };
     });
 
@@ -59,8 +76,20 @@ export default function BucketDetail(props: Readonly<Props>) {
         {title: "Size", dataIndex: "size", key: "size"},
         {title: "History", dataIndex: "history", key: "history"},
         {title: "Oldest Record (UTC)", dataIndex: "oldestRecord", key: "oldestRecord"},
-        {title: "Latest Record (UTC)", dataIndex: "latestRecord", key: "latestRecord"}
+        {title: "Latest Record (UTC)", dataIndex: "latestRecord", key: "latestRecord"},
+        {
+            title: "",
+            render: (_: any, entry: { name: string; }) => {
+                if (props.permissions?.fullAccess || (props.permissions?.write && info && props.permissions?.write?.indexOf(info?.name) !== -1)) {
+                    return <DeleteOutlined key={entry.name} style={{color: "red"}} title={"Remove entry"}
+                                           onClick={() => setEntryToRemove(entry.name)}
+                    />;
+                }
+                return <div/>;
+            }
+        }
     ];
+
 
     return <div style={{margin: "1.4em"}}>
         {info ? <BucketCard bucketInfo={info} index={0} {...props}
@@ -70,5 +99,8 @@ export default function BucketDetail(props: Readonly<Props>) {
 
         <Typography.Title level={3}>Entries</Typography.Title>
         <Table style={{margin: "0.6em"}} columns={columns} dataSource={data} loading={entries.length == 0}/>
+        <RemoveConfirmationByName name={entryToRemove} onRemoved={() => removeEntry(entryToRemove)}
+                                  onCanceled={() => setEntryToRemove("")} resourceType="entry"
+                                  confirm={entryToRemove !== ""}/>
     </div>;
 }